requirements.txt
# install these using 'python -m pip install -r requirements.txt' 

discord.py
python-dotenv
discord-py-slash-command
requests
python-dateutil